Kjy/UI/shop UI/2

+++ Assets/Scripts/Datas/Passive_Value.cs
@@ -4,8 +4,8 @@ public class Passive_Value : Data
 {
     public string ID { get; }
     public List<float> Status_UP { get; }
-    public float Status_1to2_NeedResource { get; }
-    public float Status_2to3_NeedResource { get; }
+    public int Status_1to2_NeedResource { get; }
+    public int Status_2to3_NeedResource { get; }
 
     public Passive_Value() : this("G101", new List<float>{10, 20, 30}, 100, 1000)
     {
@@ -14,8 +14,8 @@ public class Passive_Value : Data
     public Passive_Value(
         string id,
         List<float> status_Up,
-        float status_1to2_NeedResource,
-        float status_2to3_NeedResource)
+        int status_1to2_NeedResource,
+        int status_2to3_NeedResource)
     {
         ID = id;
         Status_UP = status_Up;
